# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Garage SaaS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /track/{token}:
    get:
      operationId: trackOrder
      tags: [orders]
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Tracking info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrackingInfo"

  # ── DASHBOARD ────────────────────────────────────────────────────────────────
...
  /orders:
    get:
      operationId: getOrders
      tags: [orders]
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: technicianId
          in: query
          schema:
            type: integer
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
        - name: priority
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Orders list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderList"

    post:
      operationId: createOrder
      tags: [orders]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderDetail"

    put:
      operationId: updateOrder
      tags: [orders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Status updated
  /orders/{id}/services:
    post:
      operationId: addOrderService
      tags: [orders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderServiceInput"
      responses:
        "201":
          description: Service added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderService"
  /orders/{id}/services/{serviceId}:
    delete:
      operationId: removeOrderService
      tags: [orders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: serviceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Removed
  /orders/{id}/parts:
    post:
      operationId: addOrderPart
      tags: [orders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderPartInput"
      responses:
        "201":
          description: Part added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderPart"
  /orders/{id}/parts/{partId}:
    delete:
      operationId: removeOrderPart
      tags: [orders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: partId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Removed
  /orders/{id}/invoice:
    post:
      operationId: issueInvoice
      tags: [orders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Invoice issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /orders/{id}/tracking-link:
    get:
      operationId: getOrderTrackingLink
      tags: [orders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tracking link
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrackingLink"

  # ── INVOICES ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Order:
      type: object
      required: [id, orderNumber, vehicleId, customerId, status]
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        vehicleId:
          type: integer
        customerId:
          type: integer
        status:
          type: string
        priority:
          type: string
        customerName:
          type: string
        vehiclePlate:
          type: string
        vehicleMake:
          type: string
        vehicleModel:
          type: string
        vehicleYear:
          type: integer
        estimatedCostMin:
          type: ["number", "null"]
        estimatedCostMax:
          type: ["number", "null"]
        grandTotal:
          type: ["number", "null"]
        receivedAt:
          type: string
        estimatedDeliveryAt:
          type: ["string", "null"]
        actualDeliveryAt:
          type: ["string", "null"]
        assignedTechnicianId:
          type: ["integer", "null"]
        assignedTechnicianName:
          type: ["string", "null"]
        paymentStatus:
          type: string
        customerComplaint:
          type: ["string", "null"]
        createdAt:
          type: string
    OrderList:
      type: object
      required: [data, total]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
        page:
          type: integer
    OrderDetail:
      allOf:
        - $ref: "#/components/schemas/Order"
        - type: object
          properties:
            services:
              type: array
              items:
                $ref: "#/components/schemas/OrderService"
            parts:
              type: array
              items:
                $ref: "#/components/schemas/OrderPart"
            photos:
              type: array
              items:
                $ref: "#/components/schemas/OrderPhoto"
            statusHistory:
              type: array
              items:
                $ref: "#/components/schemas/StatusHistoryItem"
            invoice:
              $ref: "#/components/schemas/Invoice"
            mileageAtReception:
              type: ["integer", "null"]
            fuelLevel:
              type: ["string", "null"]
            diagnosisNotes:
              type: ["string", "null"]
            workDescription:
              type: ["string", "null"]
            internalNotes:
              type: ["string", "null"]
            customerApproved:
              type: ["boolean", "null"]
            trackingToken:
              type: ["string", "null"]
            totalPartsCost:
              type: ["number", "null"]
            totalLaborCost:
              type: ["number", "null"]
            discountAmount:
              type: ["number", "null"]
            taxRate:
              type: ["number", "null"]
    OrderInput:
      type: object
      required: [vehicleId, customerId, customerComplaint]
      properties:
        vehicleId:
          type: integer
        customerId:
          type: integer
        customerComplaint:
          type: string
        priority:
          type: string
        mileageAtReception:
          type: integer
        fuelLevel:
          type: string
        estimatedDurationHours:
          type: number
        estimatedCostMin:
          type: number
        estimatedCostMax:
          type: number
        estimatedDeliveryAt:
          type: string
        assignedTechnicianId:
          type: integer
        internalNotes:
          type: string
    OrderUpdate:
      type: object
      properties:
        customerComplaint:
          type: string
        priority:
          type: string
        diagnosisNotes:
          type: string
        workDescription:
          type: string
        internalNotes:
          type: string
        estimatedCostMin:
          type: number
        estimatedCostMax:
          type: number
        estimatedDeliveryAt:
          type: string
        assignedTechnicianId:
          type: integer
        discountAmount:
          type: number
        discountReason:
          type: string
        taxRate:
          type: number
    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
        note:
          type: string
    OrderService:
      type: object
      required: [id, orderId, serviceNameAr, quantity, unitPrice, totalPrice]
      properties:
        id:
          type: integer
        orderId:
          type: integer
        serviceNameAr:
          type: string
        serviceNameEn:
          type: ["string", "null"]
        quantity:
          type: number
        unitPrice:
          type: number
        totalPrice:
          type: number
        status:
          type: string
        warrantyDays:
          type: ["integer", "null"]
    OrderServiceInput:
      type: object
      required: [serviceNameAr, quantity, unitPrice]
      properties:
        serviceNameAr:
          type: string
        serviceNameEn:
          type: string
        descriptionAr:
          type: string
        quantity:
          type: number
        unitPrice:
          type: number
        warrantyDays:
          type: integer
    OrderPart:
      type: object
      required: [id, orderId, partNameAr, quantity, unitCost, unitPrice]
      properties:
        id:
          type: integer
        orderId:
          type: integer
        partNameAr:
          type: string
        partNameEn:
          type: ["string", "null"]
        partNumber:
          type: ["string", "null"]
        brand:
          type: ["string", "null"]
        quantity:
          type: number
        unitCost:
          type: number
        unitPrice:
          type: number
        warrantyDays:
          type: ["integer", "null"]
    OrderPartInput:
      type: object
      required: [partNameAr, quantity, unitCost, unitPrice]
      properties:
        partNameAr:
          type: string
        partNameEn:
          type: string
        partNumber:
          type: string
        brand:
          type: string
        quantity:
          type: number
        unitCost:
          type: number
        unitPrice:
          type: number
        warrantyDays:
          type: integer
    OrderPhoto:
      type: object
      required: [id, orderId, photoUrl, type]
      properties:
        id:
          type: integer
        orderId:
          type: integer
        photoUrl:
          type: string
        thumbnailUrl:
          type: ["string", "null"]
        type:
          type: string
        captionAr:
          type: ["string", "null"]
        visibleToCustomer:
          type: boolean
        createdAt:
          type: string
    StatusHistoryItem:
      type: object
      required: [id, orderId, oldStatus, newStatus, createdAt]
      properties:
        id:
          type: integer
        orderId:
          type: integer
        oldStatus:
          type: string
        newStatus:
          type: string
        changedByName:
          type: string
        note:
          type: ["string", "null"]
        createdAt:
          type: string
    TrackingLink:
      type: object
      required: [token, url, expiresAt]
      properties:
        token:
          type: string
        url:
          type: string
        expiresAt:
          type: string
    TrackingInfo:
      type: object
      required: [orderNumber, status, vehicleInfo]
      properties:
        orderNumber:
          type: string
        status:
          type: string
        vehicleInfo:
          type: string
        customerComplaint:
          type: ["string", "null"]
        estimatedDeliveryAt:
          type: ["string", "null"]
        stages:
          type: array
          items:
            type: object
            properties:
              stage:
                type: string
              completed:
                type: boolean
              label:
                type: string

    # ── INVOICE ────────────────────────────────────────────────────────────────
    Invoice:
      type: object
      required: [id, invoiceNumber, orderId, total, status]
      properties:
        id:
          type: integer
        invoiceNumber:
          type: string
        orderId:
          type: integer
        customerId:
          type: integer
        customerName:
          type: string
        vehicleInfo:
          type: string
        subtotal:
          type: number
        taxAmount:
          type: number
        discountAmount:
          type: number
        total:
          type: number
        totalPaid:
          type: number
        balanceDue:
          type: number
        currency:
          type: string
        status:
          type: string
        paymentStatus:
          type: string
        issuedAt:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        paidAt:
          type: ["string", "null"]
        notesAr:
          type: ["string", "null"]
        createdAt:
          type: string
...